:bug: Fixed access control

+++ src/use_cases/article/deleteArticle.ts
@@ -4,12 +4,16 @@ import NotAuthorizedError from "../../types/errors/notAuthorizedError";
 import DeleteArticleRequest from "../../types/requests/article/deleteArticleRequest";
 
 export default async (request: DeleteArticleRequest) => {
-    const article = await Article.findById(request.id);
-    const restaurant = await Restaurant.findById(article.restaurantId);
-    if (request.profileId != restaurant.profileId) {
-        throw new NotAuthorizedError();
-    }
+  const article = await Article.findById(request.id);
+  const restaurant = await Restaurant.findById(article.restaurantId);
 
-    await article.delete();
+  if (!restaurant) {
+    throw new Error("Restaurant not found");
+  }
 
+  if (request.profileId != restaurant.profileId) {
+    throw new NotAuthorizedError();
+  }
+
+  await article.delete();
 };
