+++ src/use_cases/category/deleteCategory.ts
@@ -4,10 +4,15 @@ import NotAuthorizedError from "../../types/errors/notAuthorizedError";
 import DeleteCategoryRequest from "../../types/requests/category/deleteCategoryRequest";
 
 export default async (request: DeleteCategoryRequest) => {
-    const category = Category.findById(request.id);
-    const restaurant = Restaurant.findById(category.restaurantId);
-    if (request.profileId != restaurant.profileId) {
-        throw new NotAuthorizedError();
-    }
-    await category.delete();
-}
+  const category = await Category.findById(request.id);
+  const restaurant = await Restaurant.findById(category.restaurantId);
+
+  if (!restaurant) {
+    throw new Error("Restaurant not found");
+  }
+
+  if (request.profileId != restaurant.profileId) {
+    throw new NotAuthorizedError();
+  }
+  await category.delete();
+};
